Factor KafkaW Consumer and producer in separate files
